feat(memory): persist RetrievalAbstention as absence-memory entries
This change implements a robust system for tracking and persisting
retrieval abstentions (queries that found no results above threshold).
By storing these as "absence-memory" entries, the system can avoid
futile re-queries and surface knowledge gaps to operators.

Key changes:
- Introduced `HybridResult` enum and `RetrievalAbstention` struct for
  explicit failure handling.
- Implemented `AbsenceEntry` and `AbsenceStore` trait for persistence.
- Added Schema v10 migration for `csm_absences` table with upsert logic.
- Improved ID derivation using `DefaultHasher` for stable identity.
- Surfaced actual `best_score_seen` from retrieval stats to abstention
  records for better calibration.
- Added `memory_list_gaps` MCP tool to audit missing concepts.
- Integrated abstention tracking into framework text probe paths.
- Synchronized VERSION file with Cargo.toml (0.3.7).
- Refactored codebase to maintain < 500 LOC per file limit:
  - Moved MCP tool execution to `src/mcp/tools.rs`.
  - Moved best-score probe logic to `src/framework_ttl.rs`.
- Fixed Clippy warnings (`float_cmp`, `significant_drop_tightening`).
- Ensured WASM and host compilation parity.

What: Implemented persistence for retrieval failures (abstentions).
Why: To implement "environment-revealed infeasibility" as a learnable
signal per Agentic Abstention (arXiv:2606.28733).
Impact: Reduces futile RAG cycles and provides an audit trail for
unresolved concepts.
